api/requirements.txt
tensorflow==2.13.0
tensorflow-serving-api==2.13.0
fastapi==0.110.2
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.3.0
matplotlib==3.8.4
numpy==1.26.4
